uber_eats_backend/requirements.txt
asgiref==3.2.10
django==3.1.12
django-cors-headers==3.5.0
djangorestframework==3.12.4
drf-yasg==1.21.7
inflection==0.5.1
packaging==21.3
pillow==8.3.2
pytz==2021.1
PyYAML==5.3.1
setuptools==50.3.2
sqlparse==0.2.4
uritemplate==3.0.1
djongo==1.3.7
